FEAT: Create Base manager class
